chore: Move extension inference tests to their own file
